ignore missing files in verify_target_images

+++ lib/review/epubmaker.rb
@@ -202,7 +202,10 @@ module ReVIEW
       @producer.contents.each do |content|
         case content.media
         when 'application/xhtml+xml'
-          File.open("#{basetmpdir}/#{content.file}") do |f|
+          unless File.exist?(File.join(basetmpdir, content.file))
+            next
+          end
+          File.open(File.join(basetmpdir, content.file)) do |f|
             REXML::Document.new(File.new(f)).each_element('//img') do |e|
               @config['epubmaker']['force_include_images'].push(e.attributes['src'])
               if e.attributes['src'] =~ /svg\Z/i
@@ -211,6 +214,9 @@ module ReVIEW
             end
           end
         when 'text/css'
+          unless File.exist?(File.join(basetmpdir, content.file))
+            next
+          end
           File.open(File.join(basetmpdir, content.file)) do |f|
             f.each_line do |l|
               l.scan(/url\((.+?)\)/) do |_m|
